Refactored Car controller class for authorization handling

+++ backend/src/main/java/no/ntnu/project/group4/webapp/controllers/CarController.java
@@ -15,6 +15,8 @@ import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
 
 import no.ntnu.project.group4.webapp.models.Car;
+import no.ntnu.project.group4.webapp.models.User;
+import no.ntnu.project.group4.webapp.services.AccessUserService;
 import no.ntnu.project.group4.webapp.services.CarService;
 
 @CrossOrigin
@@ -23,6 +25,8 @@ import no.ntnu.project.group4.webapp.services.CarService;
 public class CarController {
   @Autowired
   private CarService carService;
+  @Autowired
+  private AccessUserService userService;
 
   @GetMapping
   public Iterable<Car> getAll() {
@@ -45,16 +49,28 @@ public class CarController {
    * Adds the specified car to the database.
    * 
    * @param car The specified car
-   * @return 201 CREATED on success or 400 BAD REQUEST on error
+   * @return <p>201 CREATED on success</p>
+   *         <p>400 BAD REQUEST on error</p>
+   *         <p>401 UNAUTHORIZED if user is not authenticated</p>
+   *         <p>403 FORBIDDEN if user is authenticated but not admin</p>
    */
   @PostMapping
   public ResponseEntity<String> add(@RequestBody Car car) {
     ResponseEntity<String> response;
-    try {
-      this.carService.add(car);
-      response = new ResponseEntity<>("", HttpStatus.CREATED);
-    } catch (Exception e) {
-      response = new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+    User sessionUser = this.userService.getSessionUser();
+    if (sessionUser != null && sessionUser.isAdmin()) {
+      try {
+        this.carService.add(car);
+        response = new ResponseEntity<>("", HttpStatus.CREATED);
+      } catch (Exception e) {
+        response = new ResponseEntity<>(e.getMessage(), HttpStatus.BAD_REQUEST);
+      }
+    } else if (sessionUser == null) {
+      response = new ResponseEntity<>("Only authenticated users have access to add cars",
+                                      HttpStatus.UNAUTHORIZED);
+    } else {
+      response = new ResponseEntity<>("Only admin users have access to add cars",
+                                      HttpStatus.FORBIDDEN);
     }
     return response;
   }
@@ -63,17 +79,29 @@ public class CarController {
    * Deletes the car with the specified ID from the database.
    * 
    * @param id The specified ID
-   * @return 200 OK on success or 404 NOT FOUND on error
+   * @return <p>200 OK on success</p>
+   *         <p>404 NOT FOUND on error</p>
+   *         <p>401 UNAUTHORIZED if user is not authenticated</p>
+   *         <p>403 FORBIDDEN if user is authenticated but not admin</p>
    */
   @DeleteMapping("/{id}")
   public ResponseEntity<String> delete(@PathVariable Long id) {
     ResponseEntity<String> response;
-    Optional<Car> car = this.carService.getOne(id);
-    if (car.isPresent()) {
-      this.carService.delete(id);
-      response = new ResponseEntity<>("", HttpStatus.OK);
+    User sessionUser = this.userService.getSessionUser();
+    if (sessionUser != null && sessionUser.isAdmin()) {
+      Optional<Car> car = this.carService.getOne(id);
+      if (car.isPresent()) {
+        this.carService.delete(id);
+        response = new ResponseEntity<>("", HttpStatus.OK);
+      } else {
+        response = new ResponseEntity<>("", HttpStatus.NOT_FOUND);
+      }
+    } else if (sessionUser == null) {
+      response = new ResponseEntity<>("Only authenticated users have access to delete cars",
+                                      HttpStatus.UNAUTHORIZED);
     } else {
-      response = new ResponseEntity<>("", HttpStatus.NOT_FOUND);
+      response = new ResponseEntity<>("Only admin users have access to delete cars",
+                                      HttpStatus.FORBIDDEN);
     }
     return response;
   }
